+++ providers/msdns/powershell.go
@@ -1,4 +1,4 @@
-package activedir
+package msdns
 
 import (
 	"bytes"
