+++ providers/msdns/powershell_test.go
@@ -1,4 +1,4 @@
-package activedir
+package msdns
 
 import (
 	"strings"
